Functions

Index

MetaRange.Chunk

MetaRange.Duration

MetaRange.Env_Preferences

MetaRange.Landscape

MetaRange.Output

MetaRange.Simulation_Data

MetaRange.Simulation_Parameters

MetaRange.Simulation_Variables

MetaRange.Species

MetaRange.Traits

MetaRange.BV

MetaRange.BVNoStoch

MetaRange.DispersalNegExpFunction

MetaRange.DispersalNegExpKernel

MetaRange.DispersalSurvivalRound

MetaRange.DispersalSurvivalStoch

MetaRange.Disperse!

MetaRange.GetDisjunctChunkGroups

MetaRange.GetDispersalSurvival

MetaRange.GetReproductionModel

MetaRange.HabitatMortality

MetaRange.KernelDispersal!

MetaRange.MortalityBev

MetaRange.MortalityBevNoStoch

MetaRange.ReproductionBeverton

MetaRange.ReproductionRicker

MetaRange.ReproductionRickerAllee

MetaRange.Survive!

MetaRange.abundance_gif

MetaRange.abundance_gif

MetaRange.carry_gif

MetaRange.check_attribute_values!

MetaRange.check_environment_dir

MetaRange.check_for_nan

MetaRange.check_species_dir

MetaRange.default_run_data

MetaRange.default_run_data

MetaRange.disperse_chunk!

MetaRange.get_Env_Preferences

MetaRange.get_Simulation_Parameters

MetaRange.get_Simulation_Variables

MetaRange.get_Traits

MetaRange.get_biomass

MetaRange.get_default_LS

MetaRange.get_default_simulation_parameters

MetaRange.get_default_species

MetaRange.get_environment_dir

MetaRange.get_habitat

MetaRange.get_is_habitat

MetaRange.get_out_dir

MetaRange.get_pop_bevmort

MetaRange.get_pop_carry

MetaRange.get_pop_var

MetaRange.get_species_dir

MetaRange.get_testrun_simulation_parameters

MetaRange.image_abundances

MetaRange.image_abundances

MetaRange.image_precipitation

MetaRange.image_precipitation

MetaRange.image_restrictions

MetaRange.image_restrictions

MetaRange.image_suitability

MetaRange.image_suitability

MetaRange.image_temperature

MetaRange.image_temperature

MetaRange.init_out_dir

MetaRange.init_species_sim_vars!

MetaRange.initialize_abundances

MetaRange.initialize_output

MetaRange.make_out_dir

MetaRange.mortality_gif

MetaRange.parse_environment_parameters!

MetaRange.parse_species_datatypes!

MetaRange.plot_abundances

MetaRange.plot_abundances

MetaRange.randomize!

MetaRange.read_input

MetaRange.read_input

MetaRange.read_species_dir

MetaRange.read_ts_config

MetaRange.reproduce!

MetaRange.reproduction_gif

MetaRange.run_simulation!

MetaRange.run_simulation!

MetaRange.save_all

MetaRange.sp_sanity_checks!

MetaRange.species_default

MetaRange.suitability_gif

MetaRange.suitability_gif

MetaRange.write_config

Function Descriptions

Exported Functions

Visualization Functions

MetaRange.plot_abundances — Function

plot_abundances(SD::Simulation_Data)

plots the total abundances of a species over time

MetaRange.image_abundances — Function

image_abundances(SD::Simulation_Data, t::Int)

plots the species abundance in the landscape for a given timestep t

MetaRange.abundance_gif — Function

abundance_gif(SD::Simulation_Data, frames=2)

creates a gif for the abundance of a species in a landscape for all timesteps

MetaRange.image_suitability — Function

image_suitability(SD::Simulation_Data, t::Int)

plots the habitat suitability of a landscape for a given timestep t

MetaRange.suitability_gif — Function

suitability_gif(SD::Simulation_Data, frames=2)

creates a gif for the habitat suitability of a landscape for all timesteps

MetaRange.image_temperature — Function

image_temperature(SD::Simulation_Data, t::Int)

plots the temperature of a landscape for a given timestep t

MetaRange.image_precipitation — Function

image_precipitation(SD::Simulation_Data, t::Int)

plots the precipitation of a landscape for a given timestep t

MetaRange.image_restrictions — Function

image_restrictions(SD::Simulation_Data, t::Int)

plots the restrictions of a landscape for a given timestep t

Internal Functions

MetaRange.Chunk — Type

Chunk

Struct for saving chunk coordinates

Fields

x::Int

y::Int

MetaRange.Duration — Type

Duration

Struct for saving the length of the runtime for the entire simulation.

Fields

start_time::DateTime: time when simulation was started

end_time::DateTime: time when simulation has finished

MetaRange.Env_Preferences — Type

Env_Preferences

Struct for saving the environmental preferences of a species.

Fields

upper_limit::Float64: species parameter

lower_limit::Float64: species parameter

optimum::Float64: species parameter

MetaRange.Landscape — Type

Landscape

Struct representing the landscape/environment.

Fields

xlength::Int64: equivalent to size[2] of any Array in this struct

ylength::Int64: equivalent to size[1] of any Array in this struct

environment::Dict{String, Array{Float64, 3}}: contains all environment attributes,

addressable by name

restrictions::Array{Float64, 3}: #TODO

MetaRange.Output — Type

Output

represents the output of a simulation run for a species.

Fields

abundances: amount of individuals for each species in each timestep

habitat: habitat suitability for each species in each timestep

carry: carrying capacity for each species in each timestep

growrate: growth rate for each species in each timestep

bevmort: background mortality rate for each species in each timestep

MetaRange.Simulation_Data — Type

Simulation_Data

Struct for saving all data related to the simulation

Fields

parameters::Simulation_Parameters: simulation parameters of the experiment

landscape::Landscape: landscape of the experiment

species::Vector{Species}: a vector of all the species in the experiment

duration::Duration: saves the duration the experiment took to compute

MetaRange.Simulation_Parameters — Type

Simulation_Parameters

Struct for saving simulation parameters.

Fields

experiment_name::String: name of the experiment

config_dir::String: path to the folder containing the config_file

config_file::String: path to the config_file

output_dir::String: path to the folder where the output should be saved

species_dir::String: path to the folder containing the species definitions

environment_dir::String: path to the folder or file containing the environment tables

input_backup::Bool: if true, the input files will be copied to the output folder

env_attribute_files::Dict{String, String}: #TODO

env_restriction_files::Dict{String, String}: #TODO

env_attribute_mode::String: #TODO

env_restriction_mode::String: #TODO

attribute_restriction_blending::String: #TODO

timesteps::Int64: number of timesteps in the simulation

randomseed::Int64: seed for number generator used in the simulation

reproduction_model::String: #TODO

use_metabolic_theory::Bool: #TODO

use_stoch_allee::Bool: if true the allee effect will be stochastic

use_stoch_carry::Bool: if true the carrying capacity will be stochastic

use_stoch_num::Bool: #TODO Stochastic Survival see R code

initialize_cells::String: #TODO

MetaRange.Simulation_Variables — Type

Simulation_Variables

Simulation variables used for a species during run_simulation!"

Fields

habitat::Array{Float64, 2}: habitability of landscape cells for a species at current

simulation timestep

is_habitat::BitArray{2}: if landscape cells are habitable for a species at current

simulation timestep

future_habitat::Array{Float64}: TODO

future_is_habitat::BitArray{2}: if landscape cells are habitable for a species at next

simulation timestep

biomass::Array{Float64, 2}: biomass of a species individual at landscape cells

growrate::Array{Float64, 2}: growrate of species at landscape cells

carry::Array{Float64, 2}: carry property of species at landscape cells

allee::Array{Float64, 2}: allee property of species at landscape cells

bevmort::Array{Float64, 2}: Beverton mortaility of species at landscape cells

occurrences::Vector{CartesianIndex{2}}: list of cells where species occurs at current

timestep

offspring::Array{Float64, 2}: offspring of species at current timestep

MetaRange.Species — Type

Species

Saving all data related to a species.

Fields

species_name::String: name of the species

traits::Traits: a (Traits)[@ref] struct for the species

abundances::Array{Union{Int64, Missing}, 3}: amount of individuals of this species in

each timestep

habitat::Array{Float64, 3}: habitat suitability in each timestep

dispersal_kernel::Matrix{Float64}:#TODO

vars::Simulation_Variables:#TODO

MetaRange.Traits — Type

Traits

Traits of a species.

#TODO: Add description in fields

Fields

mass::Float64:

sd_mass::Float64:

growrate::Float64:

sd_growrate::Float64:

param_const_growrate::Union{Float64, Nothing}:

max_dispersal_dist::Int64:

max_dispersal_buffer::Int64:

mean_dispersal_dist::Int64:

allee::Float64: Allee effect counteracting negative diversity loss in small populations

sd_allee::Float64: Allee effect standard deviation

param_const_allee::Union{Float64, Nothing}:

bevmort::Float64:

sd_bevmort::Float64:

param_const_bevmort::Union{Float64, Nothing}:

carry::Float64:

sd_carry::Float64:

param_const_carry::Union{Float64, Nothing}:

env_preferences::Dict{String, Env_Preferences}:

habitat_cutoff_suitability::Float64:

MetaRange.BV — Method

BV(N::Int64, growrate::Float64, carry::Float64, mortality::Float64)

Returns the number of individuals in the next generation according to the Beverton-Holt model. Includes stochastic mortality.

MetaRange.BVNoStoch — Method

BVNoStoch(N::Int64, growrate::Float64, carry::Float64, mortality::Float64)

Returns the number of individuals in the next generation according to the Beverton-Holt model. Does not include stochastic mortality.

MetaRange.DispersalNegExpFunction — Method

DispersalNegExpFunction(alpha, r)

TODO

MetaRange.DispersalNegExpKernel — Method

DispersalNegExpKernel(Dispersalbuffer, mean_dispersal_dist)

TODO

MetaRange.DispersalSurvivalRound — Method

DispersalSurvivalRound(
Abundances::Array{Union{Missing, Int64},2},
Offspring::Array{Float64,2}, xy::Array{Int64,2},
max_dispersal_dist::Int64
)

TODO

MetaRange.DispersalSurvivalStoch — Method

DispersalSurvivalStoch(
Abundances::Array{Union{Missing, Int64},2},
Offspring::Array{Float64,2},xy::Array{Int64,2},
max_dispersal_dist::Int64
)

TODO

MetaRange.Disperse! — Method

Disperse!(
species::Vector{Species},
LS::Landscape,
groups::NTuple{4, Vector{Chunk}},
timestep::Int64,

)

Dispersal

MetaRange.GetDisjunctChunkGroups — Method

GetDisjunctChunkGroups(
max_dispersal_buffer::Int64,
size::Tuple{Int64, Int64, Int64},

)

Calculate the distinct chunk groups

MetaRange.GetDispersalSurvival — Method

GetDispersalSurvival(use_stoch_Num::Bool)

TBW

MetaRange.GetReproductionModel — Method

GetReproductionModel(reproduction_model::String)

TBW

MetaRange.HabitatMortality — Method

HabitatMortality(Abundances::Matrix{Union{Missing,Int64}}, Is_habitat::BitArray{2})

Habitat based mortality.

This function kills individuals that are in non suitable Habitat.

Abundances: array with the number of individuals in the landscape

Is_habitat: array with boolean values that indicate

which cell is habitat in the next timestep

MetaRange.KernelDispersal! — Method

KernelDispersal!(N::Int64, Offspring::Array{Float64,2}, Dispersal_kernel::Array{Float64,2})

TODO

MetaRange.MortalityBev — Method

MortalityBev(N::Int64, mortality::Float64)

Returns how many Individuals die. Includes Stochasticity

MetaRange.MortalityBevNoStoch — Method

MortalityBevNoStoch(N::Int64, mortality::Float64)

Returns how many Individuals die. No Stochasticity

MetaRange.ReproductionBeverton — Method

ReproductionBeverton(N::Int64, growrate::Float64, carry::Float64, mortality::Float64)

Returns the number of offspring in the next generation according to the Beverton-Holt model.

MetaRange.ReproductionRicker — Method

ReproductionRicker(N::Int64, growrate::Float64, carry::Float64, unused::Union{Float64,Nothing})

Returns the number of Individuals in the next generation according to the Ricker model.

MetaRange.ReproductionRickerAllee — Method

ReproductionRickerAllee(N::Int64, growrate::Float64, carry::Float64, allee::Union{Float64,Nothing})

Returns the number of Individuals in the next generation according to the Ricker model. Includes allee effects

MetaRange.Survive! — Method

Survive!(species::Vector{Species}, DispersalSurvival, t::Int64)

TBW

MetaRange.abundance_gif — Function

abundance_gif(SD::Simulation_Data, frames=2)

creates a gif for the abundance of a species in a landscape for all timesteps

MetaRange.carry_gif — Function

carry_gif(SD::Simulation_Data, frames=2)

creates a gif for the carrying capacity of a landscape for all timesteps

MetaRange.check_attribute_values! — Method

check_attribute_values!(attribute::Array{Float64}, key::String)

Sanity checks for Attribute Matrices. Will convert Celsius values to Kelvin.

MetaRange.check_environment_dir — Method

check_environment_dir(config::Simulation_Parameters)

Checks if the environment directory exists and throws an error if it doesn't.

MetaRange.check_for_nan — Method

check_for_nan(attribute::Array{Float64})

Checks for NaNs in parameter matrix

MetaRange.check_species_dir — Method

check_species_dir(config::Simulation_Parameters)

Checks if the species directory exists and throws an error if it doesn't.

MetaRange.default_run_data — Method

default_run_data()

Initializes a simple default run Simulation_Data struct.

MetaRange.disperse_chunk! — Method

disperse_chunk!(
species::Species,
occurrences::Vector{CartesianIndex{2}},
offspring::Matrix{Float64},
chunk::Chunk,
t::Int64,

)

Parallel Dispersal in a given chunk

MetaRange.get_Env_Preferences — Method

get_Env_Preferences(species::Dict, key::String)

Returns the environmental preference for the supplied trait in key with limits and optimum as an Env_Preferences object.

MetaRange.get_Simulation_Parameters — Method

get_Simulation_Parameters(config::Dict)

Returns a Simulation_Parameters object constructed from the configuration Dictionary.

MetaRange.get_Simulation_Variables — Method

get_Simulation_Variables()

Initializes an empty Simulation_Variables object. Each parameter is defined with an empty matrix eqivalent to the landscape's size.

MetaRange.get_Traits — Method

get_Traits(species::Dict)

Extracts species traits from the configuration dictionary and returns a Trait object.

MetaRange.get_biomass — Method

get_biomass(mass, sd_mass, ysize, xsize)

TBW

MetaRange.get_default_LS — Method

get_default_LS()

Creates a Default Landscape with random values and some NAs for testing.

MetaRange.get_default_simulation_parameters — Method

get_default_simulation_parameters()

Returns a Dictionary with default simulation parameters.

MetaRange.get_default_species — Method

get_default_species(LS::Landscape, SP::Simulation_Parameters)

TBW

MetaRange.get_environment_dir — Method

get_environment_dir(config::Simulation_Parameters)

Returns full path to environment folder or returns config_dir/environment when no path to the species folder is provided.

MetaRange.get_habitat — Method

get_habitat(
env_pref::Dict{String, Env_Preferences},
LS::Landscape,
attribute_mode::String,
t::Int

)

Get habitat in current timestep. Return Array{Float64, 2}

MetaRange.get_is_habitat — Method

get_is_habitat(habitat, habitat_cutoff_suitability)

TBW

MetaRange.get_out_dir — Method

get_out_dir(SP::Simulation_Parameters)

Names a new output directory for the simulation used in init_out_dir()[@ref]. This directory will only be created if backup is true or the user later saves an output into the default paths

MetaRange.get_pop_bevmort — Method

get_pop_bevmort(
traits::Traits,
LS::Landscape,
habitat::Array{Float64,2},
mass::Union{Float64,Array{Float64,2}},
use_metabolic_theory::Bool,
timestep::Int,
E::Float64,

)

TBW

MetaRange.get_pop_carry — Method

get_pop_carry(
traits::Traits,
LS::Landscape,
habitat::Array{Float64,2},
mass::Array{Float64,2},
use_metabolic_theory::Bool,
timestep::Int,
E::Float64,

)

TBW

MetaRange.get_pop_var — Method

get_pop_var(
trait::Float64,
sd_trait::Float64,
exp_trait::Float64,
param_const_trait::Union{Float64,Nothing},
traits::Traits,
LS::Landscape,
mass::Array{Float64,2},
use_metabolic_theory::Bool,
timestep::Int, E::Float64,

)

TBW

MetaRange.get_species_dir — Method

get_species_dir(config::Simulation_Parameters)

Returns full path to species folder or returns config_dir/species when no path to the species folder is provided.

MetaRange.get_testrun_simulation_parameters — Method

get_testrun_simulation_parameters()

TBW

MetaRange.image_abundances — Method

image_abundances(SD::Simulation_Data, t::Int)

plots the species abundance in the landscape for a given timestep t

MetaRange.image_precipitation — Method

image_precipitation(SD::Simulation_Data, t::Int)

plots the precipitation of a landscape for a given timestep t

MetaRange.image_restrictions — Method

image_restrictions(SD::Simulation_Data, t::Int)

plots the restrictions of a landscape for a given timestep t

MetaRange.image_suitability — Method

image_suitability(SD::Simulation_Data, t::Int)

plots the habitat suitability of a landscape for a given timestep t

MetaRange.image_temperature — Method

image_temperature(SD::Simulation_Data, t::Int)

plots the temperature of a landscape for a given timestep t

MetaRange.init_out_dir — Method

init_out_dir(SP::Simulation_Parameters)

Initializes the output directory. This is called when input_backup in the configuration file is set to true and creates a backup of the input files in the output directory.

MetaRange.init_species_sim_vars! — Method

init_species_sim_vars!(
species::Array{Species},
LS::Landscape,
parameters::Simulation_Parameters,
timestep::Int,

)

TBW

MetaRange.initialize_abundances — Method

initialize_abundances(
    SP::Simulation_Parameters,
    habitat::Array{Float64, 2},
    carry::Float64)

Initialization of Abundances.

MetaRange.initialize_output — Method

initialize_output(
SP::Simulation_Parameters,
LS::Landscape,
abundances::Array{Int64,3},
habitat::Array{Float64,3},

)

Initializes the output struct. Calculation of the first timesteps is done outside the function.

MetaRange.make_out_dir — Method

make_out_dir(out_dir::String)

create an output directory if it does not exist

MetaRange.mortality_gif — Function

mortality_gif(SD::Simulation_Data, frames=2)

creates a gif for the mortality rate of a species in a landscape for all timesteps

MetaRange.parse_environment_parameters! — Method

parse_environment_parameters!(config::Dict, input_config::Dict)

Build the Dicts containing environment attribute and restriction files.

MetaRange.parse_species_datatypes! — Method

parse_species_datatypes!(species::Dict)

Convert all Float and Integer arguments to their respective Julia types (as needed for the species)

MetaRange.plot_abundances — Method

plot_abundances(SD::Simulation_Data)

plots the total abundances of a species over time

MetaRange.randomize! — Method

randomize!(value,sd)

Takes a parameter or array of parameters and modifies it according to a lognormal distribution based on standard deviation sd

MetaRange.read_input — Method

read_input(path::String)

Reads in the configuration and associated files in the folder and returns a Simulation_Data struct.

MetaRange.read_species_dir — Method

read_species_dir(species_dir::String, LS::Landscape, SP::Simulation_Parameters)

Read species directory, calculate properties and parameters if not provided and return as a vector of Species objects.

MetaRange.read_ts_config — Method

read_ts_config(env_dir::String, ls_timeseries_config::String)

Returns the timeseries generator configuration as a Dict (no struct as it's only used once)

MetaRange.reproduce! — Method

reproduce(species, reproduction, timestep)

Reproduction function. Takes a vector of species structs, a reproduction function and a timestep and calculates the amount of species in the next timestep.

MetaRange.reproduction_gif — Function

reproduction_gif(SD::Simulation_Data, frames=2)

creates a gif for the reproduction rate of a species in a landscape for all timesteps

MetaRange.run_simulation! — Method

run_simulation!(SD::Simulation_Data)

Ecological Simulation

Main simulation function. Load, initialize and execute all data & processes that are needed.

Arguments

SD::MetaRange.Simulation_Data: MetaRange simulation data struct. Most likely produced by read_input()

Returns

SD::MetaRange.Simulation_Data: Returns the struct with later timesteps after initialisation simulated.

MetaRange.save_all — Method

save_all(SD::Simulation_Data)

saves all output variables - reproduction, mortality rate, carrying capacity, habitat suitability, abundance - in a CSV file. The format is as follows: t, x, y, value, parameter

MetaRange.sp_sanity_checks! — Method

sp_sanity_checks!(config::Dict)

Check if necessary configuration fields are missing

MetaRange.species_default — Method

species_default()

Returns a dictionary of default species

MetaRange.suitability_gif — Function

suitability_gif(SD::Simulation_Data, frames=2)

creates a gif for the habitat suitability of a landscape for all timesteps

MetaRange.write_config — Method

write_config(SD::Simulation_Data, backup_path::String)

Record the settings actually used for a simulation run and creates a config file that can be used for future replicate runs.