Functions

Index

MetaRange.Chunk

MetaRange.Env_Preferences

MetaRange.Simulation_Data

MetaRange.Simulation_Variables

MetaRange.Traits

MetaRange.BV

MetaRange.BVNoStoch

MetaRange.DispersalNegExpFunction

MetaRange.DispersalNegExpKernel

MetaRange.DispersalSurvivalRound

MetaRange.DispersalSurvivalStoch

MetaRange.Disperse!

MetaRange.GetDisjunctChunkGroups

MetaRange.GetDispersalSurvival

MetaRange.GetReproductionModel

MetaRange.HabitatMortality

MetaRange.InitializeAbundances

MetaRange.KernelDispersal!

MetaRange.MortalityBev

MetaRange.MortalityBevNoStoch

MetaRange.ReproductionBeverton

MetaRange.ReproductionRicker

MetaRange.ReproductionRickerAllee

MetaRange.Run_Simulation!

MetaRange.Run_Simulation!

MetaRange.Survive!

MetaRange.check_attribute_values!

MetaRange.check_environmentdir!

MetaRange.check_for_nan

MetaRange.check_speciesdir!

MetaRange.default_run_data

MetaRange.default_run_data

MetaRange.disperse_chunk!

MetaRange.get_Env_Preferences

MetaRange.get_Simulation_Parameters

MetaRange.get_Simulation_Variables

MetaRange.get_Traits

MetaRange.get_biomass

MetaRange.get_habitat

MetaRange.get_is_habitat

MetaRange.get_pop_bevmort

MetaRange.get_pop_carry

MetaRange.get_pop_var

MetaRange.init_species_sim_vars!

MetaRange.parse_environment_parameters!

MetaRange.parse_species_datatypes!

MetaRange.read_input

MetaRange.read_input

MetaRange.read_species_dir

MetaRange.read_ts_config

MetaRange.reproduce!

MetaRange.sp_sanity_checks!

Function Descriptions

MetaRange.Chunk — Type

Chunk

Struct for saving chunk coordinates

Fields

-x::Int -y::Int

MetaRange.Env_Preferences — Type

Env_Preferences

Struct for saving the environmental preferences of a species.

Fields

upper_limit::Float64: species parameter

lower_limit::Float64: species parameter

optimum::Float64: species parameter

MetaRange.Simulation_Data — Type

Simulation_Data

Struct for saving all data related to the simulation

Fields

parameters::Simulation_Parameters: simulation parameters of the experiment

landscape::Landscape: landscape of the experiment

species::Vector{Species}: a vector of all the species in the experiment

duration::Duration: saves the duration the experiment took to compute

MetaRange.Simulation_Variables — Type

Simulation_Variables

Simulation variables used for a species during Run_Simulation!"

Fields

-habitat::Array{Float64, 2}: habitability of landscape cells for a species at current simulation timestep -is_habitat::BitArray{2}: if landscape cells are habitable for a species at current simulation timestep -future_habitat::Array{Float64}: TODO -future_is_habitat::BitArray{2}: if landscape cells are habitable for a species at next simulation timestep -biomass::Array{Float64, 2}: biomass of a species individual at landscape cells -growrate::Array{Float64, 2}: growrate of species at landscape cells -carry::Array{Float64, 2}: carry property of species at landscape cells -allee::Array{Float64, 2}: allee property of species at landscape cells -bevmort::Array{Float64, 2}: Beverton mortaility of species at landscape cells -occurrences::Vector{CartesianIndex{2}}: list of cells where species occurs at current timestep -offspring::Array{Float64, 2}: offspring of species at current timestep

MetaRange.Traits — Type

Traits

Traits of a species.

#TODO: Add description in fields

Fields

mass::Float64:

sd_mass::Float64:

growrate::Float64:

sd_growrate::Float64:

param_const_growrate::Union{Float64, Nothing}:

max_dispersal_dist::Int64:

max_dispersal_buffer::Int64:

mean_dispersal_dist::Int64:

allee::Float64: Allee effect counteracting negative diversity loss in small populations

sd_allee::Float64: Allee effect standard deviation

param_const_allee::Union{Float64, Nothing}:

bevmort::Float64:

sd_bevmort::Float64:

param_const_bevmort::Union{Float64, Nothing}:

carry::Float64:

sd_carry::Float64:

param_const_carry::Union{Float64, Nothing}:

env_preferences::Dict{String, Env_Preferences}:

habitat_cutoff_suitability::Float64:

MetaRange.BV — Method

BV(N::Int64, growrate::Float64, carry::Float64, mortality::Float64)

Returns the number of individuals in the next generation according to the Beverton-Holt model. Includes stochastic mortality.

MetaRange.BVNoStoch — Method

BVNoStoch(N::Int64, growrate::Float64, carry::Float64, mortality::Float64)

Returns the number of individuals in the next generation according to the Beverton-Holt model. Does not include stochastic mortality.

MetaRange.DispersalNegExpFunction — Method

DispersalNegExpFunction(alpha, r)

TODO

MetaRange.DispersalNegExpKernel — Method

DispersalNegExpKernel(Dispersalbuffer, mean_dispersal_dist)

TODO

MetaRange.DispersalSurvivalRound — Method

DispersalSurvivalRound(
Abundances::Array{Union{Missing, Int64},2},
Offspring::Array{Float64,2}, xy::Array{Int64,2},
max_dispersal_dist::Int64
)

TODO

MetaRange.DispersalSurvivalStoch — Method

DispersalSurvivalStoch(
Abundances::Array{Union{Missing, Int64},2},
Offspring::Array{Float64,2},xy::Array{Int64,2},
max_dispersal_dist::Int64
)

TODO

MetaRange.Disperse! — Method

Disperse!(
species::Vector{Species},
LS::Landscape,
groups::NTuple{4, Vector{Chunk}},
timestep::Int64,

)

Dispersal

MetaRange.GetDisjunctChunkGroups — Method

GetDisjunctChunkGroups(
max_dispersal_buffer::Int64,
size::Tuple{Int64, Int64, Int64},

)

Calculate the distinct chunk groups

MetaRange.GetDispersalSurvival — Method

GetDispersalSurvival(use_stoch_Num::Bool)

TBW

MetaRange.GetReproductionModel — Method

GetReproductionModel(reproduction_model::String)

TBW

MetaRange.HabitatMortality — Method

HabitatMortality(Abundances, Is_habitat)

Habitat based mortality

This function kills individuals that are in non suitable Habitat in code{Run}

Abundances: array with the number of individuals in the landscape Is_habitat: array with boolean values that indicate which cell is habitat in the next timestep

MetaRange.InitializeAbundances — Method

InitializeAbundances(
    SP::Simulation_Parameters,
    habitat::Array{Float64, 2},
    carry::Float64)

Initialization of Abundances.

MetaRange.KernelDispersal! — Method

KernelDispersal!(N::Int64, Offspring::Array{Float64,2}, Dispersal_kernel::Array{Float64,2})

TODO

MetaRange.MortalityBev — Method

MortalityBev(N::Int64, mortality::Float64)

Returns how many Individuals die. Includes Stochasticity

MetaRange.MortalityBevNoStoch — Method

MortalityBevNoStoch(N::Int64, mortality::Float64)

Returns how many Individuals die. No Stochasticity

MetaRange.ReproductionBeverton — Method

ReproductionBeverton(N::Int64, growrate::Float64, carry::Float64, mortality::Float64)

Returns the number of offspring in the next generation according to the Beverton-Holt model.

MetaRange.ReproductionRicker — Method

ReproductionRicker(N::Int64, growrate::Float64, carry::Float64, unused::Union{Float64,Nothing})

Returns the number of Individuals in the next generation according to the Ricker model.

MetaRange.ReproductionRickerAllee — Method

ReproductionRickerAllee(N::Int64, growrate::Float64, carry::Float64, allee::Union{Float64,Nothing})

Returns the number of Individuals in the next generation according to the Ricker model. Includes allee effects

MetaRange.Run_Simulation! — Method

Run_Simulation!(SD::Simulation_Data)

Ecological Simulation

Main simulation function. Load, initialize and execute all data & processes that are needed.

Arguments

SD::MetaRange.Simulation_Data: MetaRange simulation data struct. Most likely produced by read_input()

Returns

SD::MetaRange.Simulation_Data: Returns the struct with later timesteps after initialisation simulated.

MetaRange.Survive! — Method

Survive!(species::Vector{Species}, DispersalSurvival, t::Int64)

TBW

MetaRange.check_attribute_values! — Method

check_attribute_values!(attribute::Array{Float64}, key::String)

Sanity checks for Attribute Matrices. Will convert Celsius values to Kelvin.

MetaRange.check_environmentdir! — Method

check_environmentdir!(config::Dict, config_path::String)

Check if environment directory was given and apply default path if not.

MetaRange.check_for_nan — Method

check_for_nan(attribute::Array{Float64})

Checks for NaNs in parameter matrix

MetaRange.check_speciesdir! — Method

check_speciesdir!(config::Dict, config_path::String)

Check if species directory was given and apply default path if not

MetaRange.default_run_data — Method

default_run_data()

Initializes a simple default run Simulation_Data struct.

MetaRange.disperse_chunk! — Method

disperse_chunk!(
species::Species,
occurrences::Vector{CartesianIndex{2}},
offspring::Matrix{Float64},
chunk::Chunk,
t::Int64,

)

Parallel Dispersal in a given chunk

MetaRange.get_Env_Preferences — Method

get_Env_Preferences(species::Dict, key::String)

Returns the environmental preference for the supplied trait in "key" with limits and optimum as an Env_Preferences object.

MetaRange.get_Simulation_Parameters — Method

get_Simulation_Parameters(config::Dict)

Returns a Simulation_Parameters object constructed from the configuration Dictionary.

MetaRange.get_Simulation_Variables — Method

get_Simulation_Variables()

Initializes an empty Simulation_Variables object. Each parameter is defined with an empty matrix eqivalent to the landscape's size.

MetaRange.get_Traits — Method

get_Traits(species::Dict)

Extracts species traits from the configuration dictionary and returns a Trait object.

MetaRange.get_biomass — Method

get_biomass(mass, sd_mass, ysize, xsize)

TBW

MetaRange.get_habitat — Method

get_habitat(
env_pref::Dict{String, Env_Preferences},
LS::Landscape,
attribute_mode::String,
t::Int

)

Get habitat in current timestep. Return Array{Float64, 2}

MetaRange.get_is_habitat — Method

get_is_habitat(habitat, habitat_cutoff_suitability)

TBW

MetaRange.get_pop_bevmort — Method

get_pop_bevmort(
traits::Traits,
LS::Landscape,
habitat::Array{Float64,2},
mass::Union{Float64,Array{Float64,2}},
use_metabolic_theory::Bool,
timestep::Int,
E::Float64,

)

TBW

MetaRange.get_pop_carry — Method

get_pop_carry(
traits::Traits,
LS::Landscape,
habitat::Array{Float64,2},
mass::Array{Float64,2},
use_metabolic_theory::Bool,
timestep::Int,
E::Float64,

)

TBW

MetaRange.get_pop_var — Method

get_pop_var(
trait::Float64,
sd_trait::Float64,
exp_trait::Float64,
param_const_trait::Union{Float64,Nothing},
traits::Traits,
LS::Landscape,
mass::Array{Float64,2},
use_metabolic_theory::Bool,
timestep::Int, E::Float64,

)

TBW

MetaRange.init_species_sim_vars! — Method

init_species_sim_vars!(
species::Array{Species},
LS::Landscape,
parameters::Simulation_Parameters,
timestep::Int,

)

TBW

MetaRange.parse_environment_parameters! — Method

parse_environment_parameters!(config::Dict, input_config::Dict)

Build the Dicts containing environment attribute and restriction files.

MetaRange.parse_species_datatypes! — Method

parse_species_datatypes!(species::Dict)

Convert all Float and Integer arguments to their respective Julia types (as needed for the species)

MetaRange.read_input — Method

read_input(path::String)

Reads in the .config and associated files in the folder and returns a Simulation_Data struct.

MetaRange.read_species_dir — Method

read_species_dir(species_dir::String, LS::Landscape, SP::Simulation_Parameters)

Read species directory, calculate properties and parameters if not provided and return as a vector of Species objects.

MetaRange.read_ts_config — Method

read_ts_config(env_dir::String, ls_timeseries_config::String)

Returns the timeseries generator configuration as a Dict (no struct as it's only used once)

MetaRange.reproduce! — Method

reproduce(species, reproduction, timestep)

Reproduction function. Takes a vector of species structs, a reproduction function and a timestep and calculates the amount of species in the next timestep.

MetaRange.sp_sanity_checks! — Method

sp_sanity_checks!(config::Dict)

Check if necessary configuration fields are missing